# gen/api/openapi.yaml
openapi: 3.0.3
info:
  description: "Špecifikácia REST webových služieb slúžia pre detailný opis požadovanej\
    \ funkcionality pre Semestrálny Projekt 2 \nv rámci predmetu B-VSA akademický\
    \ rok 2022/2023.\n\nPre dopyty na väčšinu služieb (okrem createStudent a createTeacher)\
    \ je potrebné byť prihlásený. Služby pracujú s Basic\nautentifikáciou. Každý š\
    tudent a učiteľ je používateľ systému, ktorý sa môže prihlásiť.\nNiektoré služ\
    by vyžadujú autorizáciu/oprávnenie pre ich dopyt. Služba, ktorá je obmedzená oprá\
    vneniami, to má uvedené\nv popise služby.\n\nAk služba má vrátiť jeden objekt\
    \ a nie je nájdený namiesto hodnoty null vráti odpoveď 404. V prípade nečakanej\
    \ chyby\nmá byť odpoveď s kódom 500 a v správe dôvod vzniku chyby (správa zo vzniknutej\
    \ Exception).\n"
  license:
    name: Unlicense license
    url: https://unlicense.org
  title: B-VSA Semestrálny Projekt 2
  version: 1.0.0
...
servers:
- url: http://localhost:8080
...
paths:
...
  /api/students/{id}:
    delete:
      description: "Vymazanie študenta podľa identifikátora. \nVymazať študenta mô\
        že iba učiteľ (t.j. prihlásený je učiteľ) alebo prihlásený študent s rovnaký\
        m identifikátorom ako v dopyte.\n"
      operationId: deleteStudent
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
          description: Objekt študenta ktorý bol vymazaný.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: "Unauthorized request. \nDopyt musí byt autorizovaný platnou\
            \ autentifikačnou schémou. Kód v odpovedi je vždy 401.\n"
          headers:
            WWW_Authenticate:
              explode: false
              schema:
                type: string
              style: simple
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: |
            Authorized user doesn't have permission for the request.
            Autentifikovaný používateľ nemá oprávnenie na dopyt. Kód v odpovedi je vždy 403.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: |
            Requested resource was not found.
            Objekt nebolo možné nájsť. Kód v odpovedi je vždy 404.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: |
            Internal server error.
            Nepredvídateľná chyba na servery pri spracovaní dopytu. Kód v odpovedi je vždy 500.
      security:
      - basicAuth: []
      summary: Delete a student by id
      tags:
      - student
      x-accepts: application/json
    get:
      description: Získanie študenta podľa identifikátora.
      operationId: getStudent
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
          description: Objekt študenta so zadaným identifikátorom. V prípade, že taký
            študent neexistuje tak je vrátená odpoveď 404.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: "Unauthorized request. \nDopyt musí byt autorizovaný platnou\
            \ autentifikačnou schémou. Kód v odpovedi je vždy 401.\n"
          headers:
            WWW_Authenticate:
              explode: false
              schema:
                type: string
              style: simple
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: |
            Requested resource was not found.
            Objekt nebolo možné nájsť. Kód v odpovedi je vždy 404.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: |
            Internal server error.
            Nepredvídateľná chyba na servery pri spracovaní dopytu. Kód v odpovedi je vždy 500.
      security:
      - basicAuth: []
      summary: Get a student by id
      tags:
      - student
      x-accepts: application/json
...
  /api/theses:
    get:
      description: Získanie zoznamu záverečných prác.
      operationId: getAllTheses
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Thesis'
                type: array
          description: Zoznam všetkých záverečných prác v systéme. Ak nie sú vypísané
            žiadne práce je vrátený prázdny zoznam.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: "Unauthorized request. \nDopyt musí byt autorizovaný platnou\
            \ autentifikačnou schémou. Kód v odpovedi je vždy 401.\n"
          headers:
            WWW_Authenticate:
              explode: false
              schema:
                type: string
              style: simple
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: |
            Requested resource was not found.
            Objekt nebolo možné nájsť. Kód v odpovedi je vždy 404.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: |
            Internal server error.
            Nepredvídateľná chyba na servery pri spracovaní dopytu. Kód v odpovedi je vždy 500.
      security:
      - basicAuth: []
      summary: Get all theses
      tags:
      - thesis
      x-accepts: application/json
    post:
      description: "Vytvorenie záverečnej práce. Vytvoriť záverečnú prácu môže iba\
        \ učiteľ. \nVytvorená záverečná práca je vytvorená pod prihláseným učiteľ\
        om.\n"
      operationId: createThesis
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Thesis'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateThesisRequest'
          description: Objekt novo vytvorenej záverečnej práce.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: "Unauthorized request. \nDopyt musí byt autorizovaný platnou\
            \ autentifikačnou schémou. Kód v odpovedi je vždy 401.\n"
          headers:
            WWW_Authenticate:
              explode: false
              schema:
                type: string
              style: simple
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: |
            Authorized user doesn't have permission for the request.
            Autentifikovaný používateľ nemá oprávnenie na dopyt. Kód v odpovedi je vždy 403.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: |
            Internal server error.
            Nepredvídateľná chyba na servery pri spracovaní dopytu. Kód v odpovedi je vždy 500.
      security:
      - basicAuth: []
      summary: Create new thesis
      tags:
      - thesis
      x-contentType: application/json
      x-accepts: application/json
  /api/theses/{id}:
    delete:
      description: |
        Vymazanie záverečnej práce podľa identifikátora.
        Vymazať prácu môže iba prihlásený učiteľ, ktorý prácu vypísal.
      operationId: deleteThesis
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Thesis'
          description: Objekt záverečnej práce, ktorý bola vymazaná.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: "Unauthorized request. \nDopyt musí byt autorizovaný platnou\
            \ autentifikačnou schémou. Kód v odpovedi je vždy 401.\n"
          headers:
            WWW_Authenticate:
              explode: false
              schema:
                type: string
              style: simple
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: |
            Authorized user doesn't have permission for the request.
            Autentifikovaný používateľ nemá oprávnenie na dopyt. Kód v odpovedi je vždy 403.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: |
            Requested resource was not found.
            Objekt nebolo možné nájsť. Kód v odpovedi je vždy 404.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: |
            Internal server error.
            Nepredvídateľná chyba na servery pri spracovaní dopytu. Kód v odpovedi je vždy 500.
      security:
      - basicAuth: []
      summary: Delete a thesis by id
      tags:
      - thesis
      x-accepts: application/json
    get:
      description: Získanie záverečnej práce podľa identifikátora.
      operationId: getThesis
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Thesis'
          description: "Objekt záverečnej práce so zadaným identifikátorom. \nV prí\
            pade, že taká záverečná práca neexistuje, tak je vrátená odpoveď 404.\n"
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: "Unauthorized request. \nDopyt musí byt autorizovaný platnou\
            \ autentifikačnou schémou. Kód v odpovedi je vždy 401.\n"
          headers:
            WWW_Authenticate:
              explode: false
              schema:
                type: string
              style: simple
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: |
            Requested resource was not found.
            Objekt nebolo možné nájsť. Kód v odpovedi je vždy 404.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: |
            Internal server error.
            Nepredvídateľná chyba na servery pri spracovaní dopytu. Kód v odpovedi je vždy 500.
      security:
      - basicAuth: []
      summary: Get a thesis by id
      tags:
      - thesis
      x-accepts: application/json
  /api/theses/{id}/assign:
    post:
      description: |
        Priradenie záverečnej práce podľa identifikátora študentovi.
        Ak je prihlásený študent v dopyte je práca priradená jemu.
        Ak je prihlásený učiteľ je práca priradená študentovi podľa identifikátora v objekte tele dopytu.
      operationId: assignThesis
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        $ref: '#/components/requestBodies/inline_object'
        content:
          application/json:
            schema:
              properties:
                studentId:
                  format: int64
                  type: integer
              type: object
        description: Objekt s identifikátorom študenta v prípade ak je prihlásený
          učiteľ. Ak je prihlásený študent telo je ignorované.
        required: false
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Thesis'
          description: Objekt záverečnej práce, ktorá bola priradená študentovi. Vrátený
            objekt má obsahovať aj objekt študenta priradeného k prácu.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: "Unauthorized request. \nDopyt musí byt autorizovaný platnou\
            \ autentifikačnou schémou. Kód v odpovedi je vždy 401.\n"
          headers:
            WWW_Authenticate:
              explode: false
              schema:
                type: string
              style: simple
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: |
            Authorized user doesn't have permission for the request.
            Autentifikovaný používateľ nemá oprávnenie na dopyt. Kód v odpovedi je vždy 403.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: |
            Requested resource was not found.
            Objekt nebolo možné nájsť. Kód v odpovedi je vždy 404.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: |
            Internal server error.
            Nepredvídateľná chyba na servery pri spracovaní dopytu. Kód v odpovedi je vždy 500.
      security:
      - basicAuth: []
      summary: Assign a thesis by id to a student
      tags:
      - thesis
      x-contentType: application/json
      x-accepts: application/json
  /api/theses/{id}/submit:
    post:
      description: |
        Odovzdanie záverečnej práce podľa identifikátora.
        Ak je prihlásený študent v dopyte musí sa zhodovať so študentom priradeného k práci.
        Ak je prihlásený učiteľ je práca odovzdaná študentom podľa identifikátora v objekte tele dopytu,
        ktorý sa musí zhodovať so študentom priradeného k prác.
      operationId: submitThesis
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        $ref: '#/components/requestBodies/inline_object_1'
        content:
          application/json:
            schema:
              properties:
                studentId:
                  format: int64
                  type: integer
              type: object
        description: Objekt s identifikátorom študenta v prípade ak je prihlásený
          učiteľ. Ak je prihlásený študent telo je ignorované.
        required: false
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Thesis'
          description: Objekt záverečnej práce, ktorá bola odovzdaná študentom.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: "Unauthorized request. \nDopyt musí byt autorizovaný platnou\
            \ autentifikačnou schémou. Kód v odpovedi je vždy 401.\n"
          headers:
            WWW_Authenticate:
              explode: false
              schema:
                type: string
              style: simple
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: |
            Authorized user doesn't have permission for the request.
            Autentifikovaný používateľ nemá oprávnenie na dopyt. Kód v odpovedi je vždy 403.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: |
            Requested resource was not found.
            Objekt nebolo možné nájsť. Kód v odpovedi je vždy 404.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: |
            Internal server error.
            Nepredvídateľná chyba na servery pri spracovaní dopytu. Kód v odpovedi je vždy 500.
      security:
      - basicAuth: []
      summary: Submit a thesis by id assigned to a student
      tags:
      - thesis
      x-contentType: application/json
      x-accepts: application/json
...
components:
  requestBodies:
    inline_object_1:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/inline_object_1'
      required: false
    inline_object:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/inline_object'
      required: false
...
  schemas:
    Student:
      description: Objekt študenta.
      example:
        year: 1
        thesis:
          registrationNumber: registrationNumber
          author:
            year: 4
            thesis: 1
            name: name
            term: 7
            id: 3
            programme: programme
            aisId: 2
            email: email
          description: description
          publishedOn: 2000-01-23
          id: 5
          title: title
          department: department
          deadline: 2000-01-23
          type: BACHELOR
          supervisor:
            password: password
            name: name
            institute: institute
            id: 2
            theses:
            - 9
            - 9
            department: department
            aisId: 7
            email: email
          status: FREE_TO_TAKE
        name: name
        term: 5
        id: 0
        programme: programme
        aisId: 6
        email: email
      properties:
        id:
          description: Id študenta. V prípade ak je použité AIS ID ako identifikátor
            študenta tak je táto hodnota rovnaká ako v atribúte aisId.
          format: int64
          type: integer
        aisId:
          description: AIS ID študenta. Zadané pri vytvorení študenta. Musí byť unikátne
            v rámci všetkých študentov.
          format: int64
          type: integer
        name:
          type: string
        email:
          format: email
          type: string
        year:
          type: integer
        term:
          type: integer
        programme:
          type: string
        thesis:
          $ref: '#/components/schemas/Thesis'
      required:
      - aisId
      - email
      - id
      type: object
...
    StudentAltResponse:
      description: Objekt študenta s id záverečnej práce namiesto objektu.
      example:
        year: 4
        thesis: 1
        name: name
        term: 7
        id: 3
        programme: programme
        aisId: 2
        email: email
      properties:
        id:
          description: Id študenta. V prípade ak je použité AIS ID ako identifikátor
            študenta tak je táto hodnota rovnaká ako v atribúte aisId.
          format: int64
          type: integer
        aisId:
          description: AIS ID študenta. Zadané pri vytvorení študenta. Musí byť unikátne
            v rámci všetkých študentov.
          format: int64
          type: integer
        name:
          type: string
        email:
          format: email
          type: string
        year:
          type: integer
        term:
          type: integer
        programme:
          type: string
        thesis:
          description: ID záverečnej práce priradené študentovi. Ak študent nemá priradenú
            prácu, tento atribút je vynechaný.
          format: int64
          type: integer
      required:
      - aisId
      - email
      - id
      type: object
...
    TeacherAltResponse:
      description: Objekt študenta s identifikátormi záverečnej práce namiesto kolekcie
        objektov.
      example:
        password: password
        name: name
        institute: institute
        id: 2
        theses:
        - 9
        - 9
        department: department
        aisId: 7
        email: email
      properties:
        id:
          description: Id učiteľa. V prípade ak je použité AIS ID ako identifikátor
            učiteľa tak je táto hodnota rovnaká ako v atribúte aisId.
          format: int64
          type: integer
        aisId:
          description: AIS ID učiteľa. Zadané pri vytvorení učiteľa. Musí byť unikátne
            v rámci všetkých učiteľov.
          format: int64
          type: integer
        name:
          type: string
        email:
          format: email
          type: string
        password:
          format: password
          type: string
        institute:
          type: string
        department:
          type: string
        theses:
          description: ID záverečných prác vytvorené učiteľom. Ak učiteľ nevytvoril
            žiadne záverečné práce je tento atribút vynechaný.
          items:
            format: int64
            type: integer
          type: array
      required:
      - aisId
      - email
      - id
      type: object
    Thesis:
      description: Objekt záverečnej práce.
      example:
        registrationNumber: registrationNumber
        author:
          year: 4
          thesis: 1
          name: name
          term: 7
          id: 3
          programme: programme
          aisId: 2
          email: email
        description: description
        publishedOn: 2000-01-23
        id: 5
        title: title
        department: department
        deadline: 2000-01-23
        type: BACHELOR
        supervisor:
          password: password
          name: name
          institute: institute
          id: 2
          theses:
          - 9
          - 9
          department: department
          aisId: 7
          email: email
        status: FREE_TO_TAKE
      properties:
        id:
          format: int64
          type: integer
        registrationNumber:
          pattern: ^FEI-\w+$
          type: string
        title:
          type: string
        description:
          type: string
        department:
          type: string
        supervisor:
          $ref: '#/components/schemas/TeacherAltResponse'
        author:
          $ref: '#/components/schemas/StudentAltResponse'
        publishedOn:
          format: date
          type: string
        deadline:
          format: date
          type: string
        type:
          enum:
          - BACHELOR
          - MASTER
          - DISSERTATION
          type: string
        status:
          enum:
          - FREE_TO_TAKE
          - IN_PROGRESS
          - SUBMITTED
          type: string
      required:
      - department
      - id
      - registrationNumber
      - supervisor
      - title
      - type
      type: object
    CreateThesisRequest:
      description: Objekt záverečnej práce pre vytvorenie novej záverečnej práce.
      example:
        registrationNumber: registrationNumber
        description: description
        title: title
        type: BACHELOR
      properties:
        registrationNumber:
          pattern: ^FEI-\w+$
          type: string
        title:
          type: string
        description:
          type: string
        type:
          enum:
          - BACHELOR
          - MASTER
          - DISSERTATION
          type: string
      required:
      - registrationNumber
      - title
      - type
      type: object
    Message:
      description: Objekt správy pre vrátenie informácii o dopyte. Využité najmä v
        prípade vrátenia odpovedi pri chybe dopytu alebo spracovania.
      properties:
        code:
          pattern: ^\d{3}$
          type: integer
        message:
          type: string
        error:
          $ref: '#/components/schemas/Message_error'
      required:
      - code
      - message
      type: object
    inline_object:
      properties:
        studentId:
          format: int64
          type: integer
      type: object
    inline_object_1:
      properties:
        studentId:
          format: int64
          type: integer
      type: object
    Message_error:
      properties:
        type:
          type: string
        trace:
          type: string
      type: object
  securitySchemes:
    basicAuth:
      scheme: basic
      type: http